feat(actions): renamed to `zustand-refined`
