+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zustand-actions",
+  "name": "zustand-refined",
   "description": "Zustand wrapper for implementing best practices",
   "version": "0.9.0",
   "keywords": [],
@@ -8,7 +8,7 @@
     "url": "https://scottrippey.github.io/"
   },
   "license": "MIT",
-  "main": "dist/zustand-actions",
+  "main": "dist/zustand-refined",
   "devDependencies": {
     "@types/react": "^19.0.10",
     "prettier": "^3.5.3",
