+++ src/examples/counter-middleware.stories.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { createStore } from "zustand";
 import { devtools } from "zustand/middleware/devtools";
 import { immer } from "zustand/middleware/immer";
-import { createGlobalState } from "../zustand-actions";
+import { createGlobalState } from "../zustand-refined";
 
 export const [useCount, counterActions] = createGlobalState({
   store: () =>
